requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.0
gspread>=6.1.0
python-dotenv>=1.0.0
pandas>=2.2.0
qrcode[pil]>=7.4.2

